Patches by OrangeTiger
Hello Israel,

I hope I'm not annoying you with all my mails, but I really like developing the tu154b :D

So I have three new patches of which I found two on the Github Repo of DestroyFX, but for me they perfectly make sense:

1) Moved the constant "fan" sound (el_loop.wav) so it's on when the electrical power is ON, Even on battery.
2) Adjusted the hydraulic pressure for both electrical hydraulic pumps.

He increased the hydraulic pressure for the electrical hydraulic pumps. I still think they are not strong enough, from what I saw on youtube videos of the real aircraft, but it is more realistic than before.

The third change I made by myself:

3)  Added a Power switch to the Transponder.

Before there was only an image of a switch and I replaced that with a real one, which is working.

So I now also have an account on Github and can discuss the changes there. Thanks for your work.

Ragards,
OrangeTiger